requirement.txt
pandas>=2.2.2
requests>=2.31.0
apache-airflow-providers-google>=10.17.0